Move `debug` tests
