+++ tests/integration/test_main.py
@@ -220,45 +220,3 @@ def test_generate_measures_dummy_tables(tmp_path, disclosure_control_enabled, ca
             births,2021-01-01,2021-12-31,1.0,1,1,female
             """
         )
-
-
-def test_debug_show(tmp_path, call_cli):
-    @function_body_as_string
-    def definition():
-        from ehrql import create_dataset, show
-        from ehrql.tables.core import patients
-
-        dataset = create_dataset()
-        year = patients.date_of_birth.year
-        show(dataset, label="Number")
-        dataset.define_population(year > 1980)
-
-    definition_path = tmp_path / "show.py"
-    definition_path.write_text(definition)
-
-    DUMMY_DATA = textwrap.dedent(
-        """\
-        patient_id,date_of_birth
-        1,1980-06-01
-        2,1985-06-01
-        """
-    )
-    dummy_tables_path = tmp_path / "dummy_tables"
-    dummy_tables_path.mkdir()
-    dummy_tables_path.joinpath("patients.csv").write_text(DUMMY_DATA)
-
-    captured = call_cli(
-        "debug",
-        definition_path,
-        "--dummy-tables",
-        dummy_tables_path,
-    )
-
-    expected = textwrap.dedent(
-        """\
-        Show line 6: Number
-        patient_id
-        -----------------
-        """
-    ).strip()
-    assert captured.err.strip() == expected
